Added high performance heap code generator
By popular demand ... specifically to optimize CS space and time
overheads to produce the input ordering of various dirty account hashes
(on top this, this is an HPC optimized implementation).  This is more
generically useful than this particular use case though.

fd_heap builds a heap out of inputs under an arbitrary element
comparison function.  But unlike standard heap algorithms (including the
prq code generator), it does not impose any requirements on how the
element storage is organized.  Instead, it works zero-copy / in-place
with whatever application element storage management already exists.  So
this can be a useful building block for doing "sort-ish" things but with
more flexibility and less overhead than standard heaps, sorts, trees,
etc.

It is useful when dealing with streaming inputs that arrive in no
particular order (and with no particular expectations about what should
arrive) but that need to be processed in a particular order and then
ideally without any auxiliary temporary allocations / memory copies /
separate sorting passes / out-of-place sorting algos / etc ...

In such cases, we can do any order-independent processing on an input
immediately when it arrives and then insert it into the heap however it
got placed in memory on arrival.  When ready to do to the order
sensitive processing, we can then spool back the inputs from the heap in
the appropriate order via the remove_min API.  (Basically a heap sort
without memory layout restrictions and that has also been overlapped
with user computations and distributed communications overheads.)

Note that, given the zero-copy / in-place nature, it is possible to have
multiple (potentially non-disjoint) heaps for the same input stream with
minimal additional costs.