Eliminate dependency on procps-ng
Some minimal cloud host setups do not install "free" utility in
procps-ng by default.  But fd_shmem_cfg used this to get the total
amount of system memory (used to size the huge and gigantic hugetlbfs
mounts).  And every time a dependency is eliminated an angel gets its
wings.

This extracts the amount of system memory from /proc/meminfo to get the
amount of system memory instead (and also traps if this does not work).

Feels like there should be a simple method to robustly, portably and
script-friendly query the amount of memory in the system and then
without needing to scrape via tools like awk the value from diagnostic
strings meant to be read by people.