Disable -Wdangling-pointer in GCC 13
Compilation fails with the following error:

```
src/util/tpool/../scratch/fd_scratch.h:176:32: error: storing the
address of local variable ‘scratch_frame’ in ‘fd_scratch_private_frame’
[-Werror=dangling-pointer=]
  176 |   fd_scratch_private_frame     = (ulong *)fmem;
```

This is due to the fact that we are storing a stack pointer to a global
variable. However there is a matching call to `fd_scratch_detach` that
ensures the pointer is cleared.

To fix this in the brutality configuration we downgrade the error into
a warning. In the regular configuration we just disable the warning.