quic: support loopback for QUIC connections
We are trying to send outgoing packets for 127.0.0.1 to the external
facing NIC rather than on the loopback device, which won't work.

This is mostly to support development and testing, although it's
conceivable someone would send transactions via. loopback in
production as well.

When sending out an XSK to lo, we need to zero a few fields of the
packet else the kernel will silently drop it, or reject it as a
martian.