poh: fix infinite forking bug
When FD becomes leader after the previous leader skipped a slot, we
start forking infinitely.

[2024-02-06T19:19:29.534327843Z INFO  solana_core::replay_stage] new fork:104 parent:99 (leader) root:68
[2024-02-06T19:19:29.928265347Z INFO  solana_core::replay_stage] new fork:105 parent:99 (leader) root:68
[2024-02-06T19:19:30.147370965Z INFO  solana_core::window_service] num addresses: 0, top packets by source: []
[2024-02-06T19:19:30.263528659Z INFO  solana_core::repair::repair_service] repair_stats: []
[2024-02-06T19:19:30.328228537Z INFO  solana_core::replay_stage] new fork:106 parent:99 (leader) root:68
[2024-02-06T19:19:30.728228859Z INFO  solana_core::replay_stage] new fork:107 parent:99 (leader) root:68
[2024-02-06T19:19:31.128223809Z INFO  solana_core::replay_stage] new fork:108 parent:99 (leader) root:68
[2024-02-06T19:19:31.528222029Z INFO  solana_core::replay_stage] new fork:109 parent:99 (leader) root:68
[2024-02-06T19:19:31.929257167Z INFO  solana_core::replay_stage] new fork:110 parent:99 (leader) root:68
[2024-02-06T19:19:32.328253981Z INFO  solana_core::replay_stage] new fork:111 parent:99 (leader) root:68
[2024-02-06T19:19:32.728275813Z INFO  solana_core::replay_stage] new fork:112 parent:99 (leader) root:68

Here, 99 was our last slot, then the other validator was supposed to be
leader for 100-103, but didn't produce any blocks. The leader detection
code works correctly for this case, so we know that for slot 104 we are
supposed to be leader, and we correctly build off 99. Slot 105 is where
the problem starts. It should be built off 104, not 99.

The reason for this is bank.is_complete() returns false for the 104
bank, so the replay stage does not believe it is possible to build off
it.

is_complete() is false, because the bank's tick_height (in Rust) is
wrong.

In this situation, the Labs validator calls register_tick a whole bunch
of times at the start of the slot for all the "missing" ticks that
happened when we weren't leader. That is part of why it has that cache
(that I think we got rid of) for ticks.

The dumb thing is that the actual value of the hash for those ticks is
ignored except for the final tick in the block, which is used to update
the recent blockhash list. My initial thought was just to call
register_tick with a dummy value to advance tick_height, but the list is
exposed in a sysvar, which means it is consensus-critical.

However, only the last 150 blocks matter for consensus, so that gives us
a good bound on the ticks we need to store. If more than 150 slots get
skipped, we can insert bad hashes for the really old slots because
they'll get overwritten before it's observable.

So basically, when we're not leader, we need to store the block boundary
ticks in a ring buffer, then when we become leader with skipped slots in
between the reset slot and the current slot, we register these, spliced
in with dummy values when it doesn't matter.

Co-authored-by: Philip Taffet <phtaffet@jumptrading.com>