netmux: copy packets to avoid corruption
A muxer that doesn't copy packets has a problem, where the mux tile
itself might get overrun, but the downstream consumer won't realize
because the sequence number that the consumer checks after reading the
fragment gets incremented against the mcache, which is different for
the consumer and the producer.

There's a few ways to fix this:

 * You can keep two sequence numbers, and have the consumer check both
   of them.
 * You can write the sequence number check into the dcache itself
   (before and at the end of) the packet.
 * You can copy the fragments.

For now we do the 3rd option, as it's the easiest to implement.