Fix GCC 14 build
GCC 14 considers the FD_FN_SENSITIVE attribute as part of the
function signature.  We were declaring functions with the attribute
and then assigning them to function pointers without it.

Since the attribute is only available on GCC 14 but we've been
using older versions, we've never noticed that the build is broken.