refactor(ghost): init, node inset, slot key, publish, print, more tests
Introduce fd_ghost_init to the API for initializing a ghost.

Make parent a required argument for node_insert, given init
will handle the special case of initializing the root.

Change the key to use slot instead of slot_hash. This leverages the
invariant that a vote will only be upserted after successful execution
by the vote program, so the bank hash is implied to be our own.

Reimplement publish (renamed from prune) to not use an additional buffer
as well as provide a generally cleaner and better commented impl.

Rework the pretty-print API to support functionality like in the demo
directly (vs. having to be done in the caller).

Add more tests relating to all the above.