quic: free server-side tls_hs when peer ACKs handshake_done
Previously, an fd_quic server retained the tls_hs object for the
lifetime of a connection.  The fd_quic client freed it as soon as
it saw a 'handshake_done' frame.

With this patch, the server now frees the tls_hs object as soon as
the client ACKs the 'handshake_done' frame sent by the server.
Technically, the server can free it as soon as it sends it.  This
can be optimized in a future patch.

Changes:
- Move 'handshake_{complete,done_send,done_ackd},hs_data_empty'
  into a bit field
- Set 'conn->handshake_done_ackd' when a handshake_done frame is
  ACKed by the client (previously, this flag was always 0)
- Document APIs that accept a NULL tls_hs object
- Remove unused function 'fd_quic_handshake_complete'