feat(eqvoc): equivocation tile
Introduce a tile and logic (fd_eqvoc) to detect equivocating shreds.

There are 2 cases to consider:
1. Equivocation for the same FEC set
2. Equivocation for overlapping FEC sets

The first case is trivially checked by making sure the signatures of
all shreds in the same FEC set are the same.

The second case is trickier. Two (or more) FEC sets might overlap in
their shred indices. Because there are two separate FEC sets, we expect
the signatures to be different. However, this still implies
equivocation, because in order for FEC sets to overlap the same shred
idx occurs twice. See fd_eqvoc.h for details.

The tile is currently mostly a stub to setup the links and topo. The
full impl will come with subsequent PRs.