ballet, flamenco: bump offline replay instruction cnt limit to 128
We have transactions on devnet offline replay that contain more than 64
instructions. We limit instruction count per transaction to 64, and our
transaction parser enforces this limit. Bumping it to 128 everywhere has
implications for the size of fd_txn_t and the number of instructions we can
pack per microblock. So this commit just bumps it for offline replay only. It's
a less than perfect but minimal change. We will also potentially have buffer
overflow as a result of this bump, but again only in the case of offline
replay.