dedup: fix inter-tile oob read
Reported via the bounty by @eternal_Zellic. This really is not a security bug because the oob read can only be triggered by a compromised tile, and this oob read does not provide any primitive useful towards achieving inter-tile RCE. Still it is could hygiene to move the check before the usage.