dump: output wallclock timestamps in realtime mode
This is not perfect, as some tiles do not populate tspub at all, or use
some other value than fd_tickcount(). Additionally, if the original
tspub was taken more than ~2s away from when the dump tool reads it, the
timestamp decompression will return an incorrect result. Regardless,
this should be more useful than putting the mcache seq in the timestamp
field and yields the expected result for the vast majority of frags.