ethtool: disable RX-UDP-GRO-FWD instead of GRO
In #2437 GRO was disabled to solve an issue with AF_XDP and GRO, namely
the fact that GRO was creating "superpackets" out of aggregated small
UDP packets.  AF_XDP was then delivering these to the socket which was
breaking QUIC.

As it turns out, we only need to disable rx-udp-gro-forwarding to fix
this, not the entire generic-receive-offload feature. This PR swaps
the disable logic for the more specific feature.

Wonderfully, this means we can keep the overarching GRO feature enabled.
This can have a massive impact on throughput for sockets-based TCP flows,
such as the HTTP snapshot download.  An increase from ~400MBps to ~900MBps
was observed :rocket:.  This PR just warns the operator if the feature
is disabled but does not try to enable it.